Use tmux variable to indicate running Vim
This uses @tmux_navigator to hold a list of tmux panes when
vim-tmux-navigator is running.
It uses VimSuspend/VimResume (available on Neovim) to remove itself when
it is suspended (`Ctrl-z`).

(It could easily get extended to detect if vim-tmux-navigator is not
responding (e.g. when a more-prompt is active), but that would involve
an additional (async) command every time vim-tmux-navigator is used.)